refactor(types): move SystemInfo to origin-types
